Put back tests in test_jobqueue_core.py and add changelog entry.
